# openapi.json
{"openapi":"3.1.0","info":{"title":"My Products API","description":"Get the details for all products on our website","termsOfService":"https://github.com/sabouf1/FastAPI-sample-API-","contact":{"Developer Name":"Said B.","website":"https://github.com/sabouf1/FastAPI-sample-API-","Email":"said.boufares.@aol.com"},"version":"0.1.0"},"paths":{"/product":{"post":{"tags":["Products"],"summary":"Add Product","operationId":"add_product_product_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/product/{product_id}":{"get":{"tags":["Products"],"summary":"Get Product","operationId":"get_product_product__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products":{"get":{"tags":["Products"],"summary":"Get Products","operationId":"get_products_products_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Product"},"type":"array","title":"Response Get Products Products Get"}}}}}}},"/seller":{"post":{"tags":["Sellers"],"summary":"Add Seller","operationId":"add_seller_seller_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellerDisplay"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/{seller_id}":{"get":{"tags":["Sellers"],"summary":"Get Seller","operationId":"get_seller_seller__seller_id__get","parameters":[{"name":"seller_id","in":"path","required":true,"schema":{"type":"integer","title":"Seller Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Seller"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Product":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"price":{"type":"integer","title":"Price"},"id":{"type":"integer","title":"Id"},"seller":{"$ref":"#/components/schemas/SellerDisplay"}},"type":"object","required":["name","description","price","id","seller"],"title":"Product"},"Seller":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["username","email","password","id"],"title":"Seller"},"SellerCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"SellerCreate"},"SellerDisplay":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"}},"type":"object","required":["username","email"],"title":"SellerDisplay"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}